+++ android/app/src/main/res/raw/keep.xml
@@ -1,2 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools" tools:keep="@drawable/ic_notification" />
+
+
+
+
+
+
