+++ lib/views/settings/settings_home_select.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:points_verts/services/mapbox.dart';
+import 'package:points_verts/services/map/map_interface.dart';
+import 'package:points_verts/services/map/mapbox.dart';
 import 'package:points_verts/models/address_suggestion.dart';
 
 import '../loading.dart';
@@ -22,6 +23,8 @@ class SettingsHomeSelect extends StatefulWidget {
 
 class _SettingsHomeSelectState extends State<SettingsHomeSelect> {
   _SettingsHomeSelectState(this.setHomeCallback, this.removeHomeCallback);
+
+  final MapInterface map = new MapBox();
 
   final Function(AddressSuggestion) setHomeCallback;
   final Function removeHomeCallback;
@@ -47,7 +50,7 @@ class _SettingsHomeSelectState extends State<SettingsHomeSelect> {
     _debounce = Timer(const Duration(milliseconds: 500), () {
       if (!mounted) return;
       setState(() {
-        _suggestions = retrieveSuggestions(
+        _suggestions = map.retrieveSuggestions(
             countryCodes[_countryIndex], _homeSearchController.text);
       });
     });
